+++ modules/dublincore.go
@@ -1,4 +1,4 @@
-package module
+package modules
 
 import ()
